adding layer to config
