+++ src/Dotknet.RegistryClient/Models/Descriptor.cs
@@ -8,16 +8,22 @@ public class Descriptor
 {
   [JsonPropertyName("mediaType")]
   public MediaTypeEnum MediaType { get; set; }
+
   [JsonPropertyName("size")]
   public long Size { get; set; }
+
   [JsonPropertyName("digest")]
   public Hash Digest { get; set; }
+
   [JsonPropertyName("data")]
   public byte[]? Data { get; set; }
+
   [JsonPropertyName("urls")]
   public IEnumerable<string>? Urls { get; set; }
+
   [JsonPropertyName("annotations")]
   public Dictionary<string, string>? Annotations { get; set; }
+  
   [JsonPropertyName("platform")]
   public Platform? Platform { get; set; }
 }
