+++ src/Dotknet.RegistryClient/Models/Hash.cs
@@ -44,6 +44,7 @@ public class HashConverter : JsonConverter<Hash>
 
   public override void Write(Utf8JsonWriter writer, Hash value, JsonSerializerOptions options)
   {
+    // todo
     writer.WriteRawValue(value.ToString());
   }
 }
